Cut underwater objects
NavMesh covers all ground and water sufraces. Bottom also covered with
NavMesh. Recastnavigation builds path by the surface, so even swiming
actor will have either water surface path or bottom surface path. And
there is a gap between ground and bottom. When actor is close to shore
and need to go on the ground recastnavigation put start path positon on
the bottom not on the water surface, where the gap prevents to find all
path points to target. Using dtCrowd leads to almost impossible for
actor to go through this gap.

To break this problem, we can cut all underwater NavMesh. So there is no
choice whether to use water surface or bottom but only water surface. Of
course there are some drawbacks. If there is an actor who moves only
over the bottom (couldn't swim but could breathe underwater), it could
not find correct path now. If bottom is too far from surface it couldn't
find any path.