Change job dispatching logic
Previously we used fixed sized batch to allocate work to each thread, so
that the movement were (almost) evenly distributed among threads.
It means that if a thread is unfortunate to have "a lot" of "difficult"
movement to solve, the other ones would wait for it, wasting cycles.

Instead, use a "first come, first serve" principle to dispatch the movements one
by one. This way, the workload is evenly distributed among threads.